Refactor: Pull transforms and validations into their own modules
