+++ src/lib.ts
@@ -2,8 +2,6 @@ import {
   buildASTSchema,
   DocumentNode,
   GraphQLSchema,
-  isAbstractType,
-  isType,
   Kind,
   validateSchema,
 } from "graphql";
@@ -14,17 +12,24 @@ import {
   DiagnosticsResult,
   Result,
   ReportableDiagnostics,
-  gqlErr,
   combineResults,
-  collectResults,
 } from "./utils/DiagnosticError";
 import * as ts from "typescript";
-import { ExtractionSnapshot, extract } from "./Extractor";
-import { GratsDefinitionNode, TypeContext } from "./TypeContext";
+import { ExtractionSnapshot } from "./Extractor";
+import { TypeContext } from "./TypeContext";
 import { validateSDL } from "graphql/validation/validate";
+import { ParsedCommandLineGrats } from "./gratsConfig";
+import { validateTypenames } from "./validations/validateTypenames";
+import { snapshotsFromProgram } from "./transforms/snapshotsFromProgram";
+import { validateMergedInterfaces } from "./validations/validateMergedInterfaces";
+import { validateContextReferences } from "./validations/validateContextReferences";
+import { GratsDefinitionNode } from "./GraphQLConstructor";
 import { DIRECTIVES_AST } from "./metadataDirectives";
 import { extend } from "./utils/helpers";
-import { ParsedCommandLineGrats } from "./gratsConfig";
+import { addInterfaceFields } from "./transforms/addInterfaceFields";
+import { filterNonGqlInterfaces } from "./transforms/filterNonGqlInterfaces";
+import { resolveTypes } from "./transforms/resolveTypes";
+import { validateAsyncIterable } from "./validations/validateAsyncIterable";
 
 export * from "./gratsConfig";
 
@@ -61,60 +66,30 @@ function extractSchema(
 ): DiagnosticsResult<GraphQLSchema> {
   const program = ts.createProgram(options.fileNames, options.options, host);
 
-  const errors: ts.Diagnostic[] = [];
-  const gratsSourceFiles = program.getSourceFiles().filter((sourceFile) => {
-    // If the file doesn't contain any GraphQL definitions, skip it.
-    if (!/@gql/i.test(sourceFile.text)) {
-      return false;
-    }
-
-    if (options.raw.grats.reportTypeScriptTypeErrors) {
-      // If the user asked for us to report TypeScript errors, then we'll report them.
-      const typeErrors = ts.getPreEmitDiagnostics(program, sourceFile);
-      if (typeErrors.length > 0) {
-        extend(errors, typeErrors);
-        return false;
-      }
-    } else {
-      // Otherwise, we will only report syntax errors, since they will prevent us from
-      // extracting any GraphQL definitions.
-      const syntaxErrors = program.getSyntacticDiagnostics(sourceFile);
-      if (syntaxErrors.length > 0) {
-        // It's not very helpful to report multiple syntax errors, so just report
-        // the first one.
-        errors.push(syntaxErrors[0]);
-        return false;
-      }
-    }
-    return true;
-  });
-
-  const extractResults = gratsSourceFiles.map((sourceFile) => {
-    return extract(options.raw.grats, sourceFile);
-  });
-
-  const snapshotResults = collectResults(extractResults);
-  if (snapshotResults.kind === "ERROR") {
-    extend(errors, snapshotResults.err);
-    return err(errors); // Needed to satisfy type checker
+  const snapshotsResult = snapshotsFromProgram(program, options);
+  if (snapshotsResult.kind === "ERROR") {
+    return snapshotsResult;
   }
 
-  if (errors.length > 0) {
-    return err(errors);
-  }
+  const snapshot = reduceSnapshots(snapshotsResult.value);
 
-  const snapshot = reduceSnapshots(snapshotResults.value);
-
-  const checker = program.getTypeChecker();
-
-  const docResult = docFromSnapshot(options, checker, host, snapshot);
+  const docResult = docFromSnapshot(program, host, snapshot);
 
   if (docResult.kind === "ERROR") {
-    return err(docResult.err);
+    return docResult;
   }
 
-  const doc = docResult.value;
+  return buildSchemaFromDocumentNode(
+    docResult.value,
+    snapshot.typesWithTypenameField,
+  );
+}
 
+// Given a SDL AST, build and validate a GraphQLSchema.
+function buildSchemaFromDocumentNode(
+  doc: DocumentNode,
+  typesWithTypenameField: Set<string>,
+): DiagnosticsResult<GraphQLSchema> {
   // TODO: Currently this does not detect definitions that shadow builtins
   // (`String`, `Int`, etc). However, if we pass a second param (extending an
   // existing schema) we do! So, we should find a way to validate that we don't
@@ -136,65 +111,34 @@ function extractSchema(
     return err(diagnostics);
   }
 
-  const typenameDiagnostics = validateTypename(
-    schema,
-    snapshot.typesWithTypenameField,
-  );
+  const typenameDiagnostics = validateTypenames(schema, typesWithTypenameField);
   if (typenameDiagnostics.length > 0) return err(typenameDiagnostics);
 
   return ok(schema);
 }
 
-function validateTypename(
-  schema: GraphQLSchema,
-  hasTypename: Set<string>,
-): ts.Diagnostic[] {
-  const typenameDiagnostics: ts.Diagnostic[] = [];
-  const abstractTypes = Object.values(schema.getTypeMap()).filter(
-    isAbstractType,
-  );
-  for (const type of abstractTypes) {
-    // TODO: If we already implement resolveType, we don't need to check implementors
-
-    const typeImplementors = schema.getPossibleTypes(type).filter(isType);
-    for (const implementor of typeImplementors) {
-      if (!hasTypename.has(implementor.name)) {
-        const loc = implementor.astNode?.name?.loc;
-        if (loc == null) {
-          throw new Error(
-            `Grats expected the parsed type \`${implementor.name}\` to have location information. This is a bug in Grats. Please report it.`,
-          );
-        }
-        typenameDiagnostics.push(
-          gqlErr(
-            loc,
-            `Missing __typename on \`${implementor.name}\`. The type \`${type.name}\` is used in a union or interface, so it must have a \`__typename\` field.`,
-          ),
-        );
-      }
-    }
-  }
-  return typenameDiagnostics;
-}
-
+/**
+ * Given a merged snapshot representing the whole program, construct a GraphQL
+ * schema document with metadata directives attached.
+ */
 export function docFromSnapshot(
-  options: ts.ParsedCommandLine,
-  checker: ts.TypeChecker,
+  program: ts.Program,
   host: ts.CompilerHost,
   snapshot: ExtractionSnapshot,
 ): DiagnosticsResult<DocumentNode> {
-  const ctx = new TypeContext(options, checker, host);
+  const checker = program.getTypeChecker();
+  const ctx = new TypeContext(checker, host);
 
-  // Phase 1: Validate the snapshot
+  // Validate the snapshot
   const mergedResult = combineResults(
-    ctx.validateMergedInterfaces(snapshot.interfaceDeclarations),
-    ctx.validateContextReferences(snapshot.contextReferences),
+    validateMergedInterfaces(checker, snapshot.interfaceDeclarations),
+    validateContextReferences(ctx, snapshot.contextReferences),
   );
   if (mergedResult.kind === "ERROR") {
     return mergedResult;
   }
 
-  // Phase 2: Propagate snapshot data to type context
+  // Propagate snapshot data to type context
 
   for (const [node, typeName] of snapshot.unresolvedNames) {
     ctx.markUnresolvedType(node, typeName);
@@ -204,7 +148,7 @@ export function docFromSnapshot(
     ctx.recordTypeName(node, definition.name, definition.kind);
   }
 
-  // Phase 3: Fixup the schema SDL
+  // Fixup the schema SDL
 
   const definitions: GratsDefinitionNode[] = Array.from(
     DIRECTIVES_AST.definitions,
@@ -215,20 +159,22 @@ export function docFromSnapshot(
   // If you define a field on an interface using the functional style, we need to add
   // that field to each concrete type as well. This must be done after all types are created,
   // but before we validate the schema.
-  const definitionsResult = ctx.handleAbstractDefinitions(definitions);
+  const definitionsResult = addInterfaceFields(ctx, definitions);
   if (definitionsResult.kind === "ERROR") {
     return definitionsResult;
   }
 
-  const docResult = ctx.resolveTypes({
+  const filteredDoc = filterNonGqlInterfaces(ctx, {
     kind: Kind.DOCUMENT,
     definitions: definitionsResult.value,
   });
+
+  const docResult = resolveTypes(ctx, filteredDoc);
   if (docResult.kind === "ERROR") return docResult;
 
   const doc = docResult.value;
 
-  const subscriptionsValidationResult = ctx.validateAsyncIterableFields(doc);
+  const subscriptionsValidationResult = validateAsyncIterable(doc);
   if (subscriptionsValidationResult.kind === "ERROR") {
     return subscriptionsValidationResult;
   }
